This is the second step in refactoring the libsrc4 code. The first was branch newhash0.dmh.
As with newhash0.dmh, these changes should be transparent.

+++ ncdap_test/test_nstride_cached.c
@@ -149,7 +149,7 @@ main()
 #ifdef VERBOSE
     for (idim=0; idim<ndim; idim++)
 	printf("start[%1d]=%3lu count[%1d]=%3lu stride[%1d]=%3lu\n",
-		idim,start[idim],idim,count[idim],idim,stride[idim]);
+		idim,(unsigned long)start[idim],idim,(unsigned long)count[idim],idim,(unsigned long)stride[idim]);
 #endif
 
     ncstatus = nc_get_vars_float (ncid, varid, start, count, stride, (float*) dat);
@@ -199,7 +199,7 @@ main()
 #ifdef VERBOSE
     for (idim=0; idim<ndim; idim++)
         printf("start[%d]=%3lu count[%d]=%3lu stride[%d]=%3lu\n",
-		idim, start[idim], idim, count[idim], idim, stride[idim]);
+		idim, (unsigned long)start[idim], idim, (unsigned long)count[idim], idim, (unsigned long)stride[idim]);
 #endif
 
     ncstatus = nc_get_vars_float (ncid, varid, start, count, stride, (float*) dat);
@@ -256,7 +256,7 @@ main()
 #ifdef VERBOSE
     for (idim=0; idim<ndim; idim++)
 	printf("start[%1d]=%3lu count[%1d]=%3lu stride[%1d]=%3lu\n",
-		idim,start[idim],idim,count[idim],idim,stride[idim]);
+		idim,(unsigned long)start[idim],idim,(unsigned long)count[idim],idim,(unsigned long)stride[idim]);
 #endif
 
     memset((void*)sdat,0,sizeof(sdat));
